requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
supabase==2.3.4
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
cloudinary==1.36.0
schedule==1.2.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
